duplications fix
